Moves ColumnList into the components folder
