+++ amundsen_application/static/js/components/ColumnList/ColumnDescEditableText/index.tsx
