+++ amundsen_application/static/js/components/ColumnList/ColumnType/styles.scss
